requirements.txt
asgiref==3.4.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.4
Django==3.2.13
django-cors-headers==3.10.1
djangorestframework==3.13.1
idna==3.3
passlib==1.7.4
Pillow==8.4.0
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
qrcode==7.3.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
twilio==7.9.2
typing_extensions==4.1.1
urllib3==1.26.9
xlwt==1.3.0
